requirements.txt
asttokens==2.4.1
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
fonttools==4.46.0
idna==3.6
ipython==8.18.1
jedi==0.19.1
kiwisolver==1.4.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
numpy==1.26.2
packaging==23.2
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
ping3==4.0.4
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pythonping==1.1.4
requests==2.31.0
six==1.16.0
stack-data==0.6.3
traitlets==5.14.0
urllib3==2.1.0
wcwidth==0.2.12
